Refactor SSO enforcement tests and clean up implementation
Consolidated SSO enforcement tests into db.spec.ts following naming conventions. Removed redundant test file and cleaned up bulkCreate implementation.
